W8/requirements.txt
faker==22.0.0
matplotlib==3.8.2
pillow==10.2.0
numpy==1.26.3

